Add Level-10 v2
