+++ src/main/java/duke/Duke.java
@@ -2,7 +2,6 @@ package duke;
 
 import java.time.format.DateTimeParseException;
 import java.util.List;
-import java.util.Scanner;
 
 import duke.exceptions.DukeException;
 import duke.exceptions.InvalidInputException;
@@ -52,9 +51,9 @@ public class Duke {
 
             case LIST:
                 if (tasks.isEmpty()) {
-                    output =  ui.showNoTasks();
+                    output = ui.showNoTasks();
                 } else {
-                    output =  ui.showTasks(tasks.getTasksDes(1), 0);
+                    output = ui.showTasks(tasks.getTasksDes(1), 0);
                 }
                 break;
 
@@ -62,7 +61,7 @@ public class Duke {
             case UNMARK:
             case MARK:
                 String completionStatus = tasks.changeTaskCompletion(input, command);
-                output =  ui.showStatusChanged(completionStatus);
+                output = ui.showStatusChanged(completionStatus);
                 break;
 
             case FIND:
@@ -94,81 +93,4 @@ public class Duke {
 
         return output;
     }
-
-
-
-
-    /**
-     * Starts the chatbot.
-     */
-    public void run() {
-
-        boolean end = false;
-
-        Scanner scanner = new Scanner(System.in);
-
-        ui.greet();
-
-        while (!end) {
-            try {
-                String nextInput = scanner.nextLine().trim();
-                Commands command = Parser.determineCommand(nextInput);
-
-                switch (command) {
-                case TODO:
-                case DEADLINE:
-                case EVENT:
-                    Task t = TaskList.createTask(nextInput, command, 0);
-                    tasks.addTask(t);
-                    ui.showTaskAdded(t.getTask());
-                    break;
-
-                case LIST:
-                    if (tasks.isEmpty()) {
-                        ui.showNoTasks();
-                    } else {
-                        ui.showTasks(tasks.getTasksDes(1), 1);
-                    }
-                    break;
-
-
-                case UNMARK:
-                case MARK:
-                    String completionStatus = tasks.changeTaskCompletion(nextInput, command);
-                    ui.showStatusChanged(completionStatus);
-                    break;
-
-                case FIND:
-                    List<String> matchingTasks = tasks.findTask(nextInput);
-                    ui.showTasks(matchingTasks, 1);
-                    break;
-
-                case DELETE:
-                    String deleteStatus = tasks.deleteTask(nextInput);
-                    ui.showStatusChanged(deleteStatus);
-                    break;
-
-                case BYE:
-                    end = true;
-                    String savedStatus = storage.saveToDisk(tasks.getTasksDes(0));
-                    ui.showStatusChanged(savedStatus);
-                    break;
-
-                case UNKNOWN:
-                default:
-                    throw new InvalidInputException("Invalid input");
-                }
-
-            } catch (DukeException e) {
-                ui.showDukeError(e);
-            } catch (DateTimeParseException e) {
-                ui.showDateError();
-            } catch (Exception e) {
-                ui.showGeneralError();
-            }
-            ui.separator();
-        }
-        ui.farewell();
-    }
-
 }
