--- src/main/java/duke/Main.java
@@ -1,105 +1,0 @@
-package duke;
-
-import java.io.IOException;
-
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Scene;
-import javafx.scene.layout.AnchorPane;
-import javafx.stage.Stage;
-
-/**
- * A GUI for Duke using FXML.
- */
-public class Main extends Application {
-
-    private Duke duke = new Duke();
-
-    @Override
-    public void start(Stage stage) {
-        try {
-            FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/view/MainWindow.fxml"));
-            AnchorPane ap = fxmlLoader.load();
-            Scene scene = new Scene(ap);
-            stage.setScene(scene);
-            fxmlLoader.<MainWindow>getController().setDuke(duke);
-            stage.show();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-//    @Override
-//    public void start(Stage stage) {
-//        //Step 1. Setting up required components
-//
-//        //The container for the content of the chat to scroll.
-//        ScrollPane scrollPane = new ScrollPane();
-//        VBox dialogContainer = new VBox();
-//        scrollPane.setContent(dialogContainer);
-//
-//        TextField userInput = new TextField();
-//        Button sendButton = new Button("Send");
-//
-//        AnchorPane mainLayout = new AnchorPane();
-//        mainLayout.getChildren().addAll(scrollPane, userInput, sendButton);
-//
-//        Scene scene = new Scene(mainLayout);
-//
-//        //Step 2. Formatting the window to look as expected
-//        stage.setTitle("Toothless");
-//        stage.setResizable(false);
-//        stage.setMinHeight(600.0);
-//        stage.setMinWidth(400.0);
-//
-//        mainLayout.setPrefSize(400.0, 600.0);
-//
-//        scrollPane.setPrefSize(385, 535);
-//        scrollPane.setHbarPolicy(ScrollPane.ScrollBarPolicy.NEVER);
-//        scrollPane.setVbarPolicy(ScrollPane.ScrollBarPolicy.ALWAYS);
-//
-//        scrollPane.setVvalue(1.0);
-//        scrollPane.setFitToWidth(true);
-//
-//        dialogContainer.setPrefHeight(Region.USE_COMPUTED_SIZE);
-//
-//        userInput.setPrefWidth(325.0);
-//
-//        sendButton.setPrefWidth(55.0);
-//
-//        AnchorPane.setTopAnchor(scrollPane, 1.0);
-//
-//        AnchorPane.setBottomAnchor(sendButton, 1.0);
-//        AnchorPane.setRightAnchor(sendButton, 1.0);
-//
-//        AnchorPane.setLeftAnchor(userInput , 1.0);
-//        AnchorPane.setBottomAnchor(userInput, 1.0);
-//
-//        //Scroll down to the end every time dialogContainer's height changes.
-//        dialogContainer.heightProperty().addListener((observable) -> scrollPane.setVvalue(1.0));
-//
-//        //Step 3. Add functionality to handle user input.
-////        sendButton.setOnMouseClicked((event) -> {
-////            dialogContainer.getChildren().add(getDialogLabel(userInput.getText()));
-////            userInput.clear();
-////        });
-////
-////        userInput.setOnAction((event) -> {
-////            dialogContainer.getChildren().add(getDialogLabel(userInput.getText()));
-////            userInput.clear();
-////        });
-//        sendButton.setOnMouseClicked((event) -> {
-//            handleUserInput(userInput, dialogContainer);
-//        });
-//
-//        userInput.setOnAction((event) -> {
-//            handleUserInput(userInput, dialogContainer);
-//        });
-//
-//
-//
-//        stage.setScene(scene);
-//        stage.show();
-//    }
-}
-
